Finalize Robot Companion: production configuration
