+++ content/components/robot-companion/gemini-service.js
@@ -2,7 +2,8 @@ import { config } from './config.js';
 
 export class GeminiService {
   constructor() {
-    this.apiKey = config.geminiApiKey;
+    // Check if key is a function (obfuscated) or string
+    this.apiKey = typeof config.getGeminiApiKey === 'function' ? config.getGeminiApiKey() : config.geminiApiKey;
     this.baseUrl = config.apiBaseUrl;
     this.model = config.model;
   }
@@ -17,7 +18,7 @@ export class GeminiService {
 
   async generateResponse(prompt, context = []) {
     if (!this.hasApiKey()) {
-      return 'Ich benötige einen API-Key für Google Gemini, um antworten zu können. Bitte konfiguriere ihn in den Einstellungen.';
+      return this.fallbackResponse(prompt);
     }
 
     const url = `${this.baseUrl}/${this.model}:generateContent?key=${this.apiKey}`;
@@ -38,6 +39,8 @@ export class GeminiService {
     try {
       const response = await fetch(url, {
         method: 'POST',
+        // Send origin/referrer safely to satisfy Google's domain check
+        referrerPolicy: 'strict-origin-when-cross-origin',
         headers: {
           'Content-Type': 'application/json',
         },
@@ -51,15 +54,24 @@ export class GeminiService {
       });
 
       if (!response.ok) {
-        const errorData = await response.json();
+        const errorData = await response.json().catch(() => ({}));
+
+        // Log detailed error for debugging (internal only)
+        if (errorData.error && errorData.error.message) {
+             console.warn('Gemini API Warning:', errorData.error.message);
+        }
 
         if (response.status === 429) {
           console.warn('Gemini API Rate Limit Hit (429)');
           return 'Ich bin gerade etwas überlastet (zu viele Anfragen). Bitte versuche es in ein paar Sekunden noch einmal. 🤯';
         }
 
-        console.error('Gemini API Error:', errorData);
-        throw new Error(`API Error: ${response.status}`);
+        if (response.status === 403) {
+             console.warn('Gemini API Key blocked or restricted.');
+             return this.fallbackResponse(prompt);
+        }
+
+        return this.fallbackResponse(prompt);
       }
 
       const data = await response.json();
@@ -67,12 +79,21 @@ export class GeminiService {
       if (data.candidates && data.candidates.length > 0 && data.candidates[0].content) {
         return data.candidates[0].content.parts[0].text;
       } else {
-        return 'Entschuldigung, ich konnte keine Antwort generieren.';
+        return this.fallbackResponse(prompt);
       }
     } catch (error) {
       console.error('Gemini Service Error:', error);
-      return 'Es gab ein Problem bei der Verbindung zu meinem Gehirn (API Error).';
+      return this.fallbackResponse(prompt);
     }
+  }
+
+  fallbackResponse(prompt) {
+      const lower = prompt.toLowerCase();
+      if (lower.includes('hallo') || lower.includes('hi')) return 'Hallo! Mein Gehirn ist gerade offline, aber ich bin trotzdem für dich da. 😊';
+      if (lower.includes('wer bist du')) return 'Ich bin dein virtueller Assistent auf dieser Webseite!';
+      if (lower.includes('hilfe')) return 'Ich kann dir helfen, dich zurechtzufinden. Klicke einfach auf die Optionen unten.';
+      if (lower.includes('witz')) return 'Warum können Geister so schlecht lügen? Weil man durch sie hindurchsehen kann! 👻 (Sorry, Offline-Modus Witz)';
+      return 'Ich habe gerade keine Verbindung zu meinem Sprachzentrum. Bitte nutze die Buttons unten für die Navigation!';
   }
 
   async summarizePage(pageContent) {
